requirements.txt
asgiref==3.4.0
dj-database-url==0.5.0
Django==3.2.4
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.2.0
pkg-resources==0.0.0
psycopg2==2.8.6
pyparsing==2.4.7
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
